:recylce: organize

+++ src/sentry/replays/usecases/ingest/event_parser.py
@@ -258,8 +258,8 @@ def get_timestamp_unit(event_type: EventType) -> Literal["s", "ms"]:
 
 def get_timestamp_ms(event: dict[str, Any], event_type: EventType) -> float:
     if get_timestamp_unit(event_type) == "s":
-        return event.get("timestamp", 0) * 1000
-    return event.get("timestamp", 0)
+        return float(event.get("timestamp", 0) * 1000)
+    return float(event.get("timestamp", 0))
 
 
 #
